testing/troubleshooting timezone issues
Looks like -0700 might not be the correct timezone in the code since publishing the post still shows up as being for the next day on the blog.  In this case, I wrote this post on September 28th but it was published under September 29th on the blog. My guess is it might have something to do with daylight savings, so changing the timezone by an hour might help get it accurate to my actual timezone (PST/PDT).